improved tests coverage. added subdirs for files
Squashed commit of the following:

commit bce52c89034c36e0b73641d0b05714b76b080cb4
Author: Paul Etchells (interceptor) <p.a.etchells@gmail.com>
Date:   Fri Feb 7 00:42:13 2025 +0100

    improved test coverage

commit d0f1fd6834cb905fa4b2df36a5bd2a70fc78d49b
Author: Paul Etchells (interceptor) <p.a.etchells@gmail.com>
Date:   Thu Feb 6 23:16:41 2025 +0100

    cleaned up tests

commit e9739b08f2c89dd6c0246b30ae854aed619bca38
Author: Paul Etchells (interceptor) <p.a.etchells@gmail.com>
Date:   Thu Feb 6 15:49:08 2025 +0100

    JSON number is always int64

commit cf3b5fa59af7103579fe2bd1d82d377ad4b4305b
Author: Paul Etchells (interceptor) <p.a.etchells@gmail.com>
Date:   Thu Feb 6 00:16:33 2025 +0100

    fixed tests

commit 55b9e460fe8e2dedd07119a32d9ac10e5e772d9b
Author: Paul Etchells (interceptor) <p.a.etchells@gmail.com>
Date:   Wed Feb 5 23:12:13 2025 +0100

    put files base/sourcename/sessID/10000/*

+++ internal/nrtm4/persist/jsondefs.go
@@ -9,18 +9,18 @@ const (
 
 // FileRefJSON json model of a file reference in a Notification file
 type FileRefJSON struct {
-	Version uint32 `json:"version"`
+	Version int64  `json:"version"`
 	URL     string `json:"url"`
 	Hash    string `json:"hash"`
 }
 
 // NrtmFileJSON json model of fields common to all NRTM4 files
 type NrtmFileJSON struct {
-	NrtmVersion uint   `json:"nrtm_version"`
+	NrtmVersion int64  `json:"nrtm_version"`
 	Type        string `json:"type"`
 	Source      string `json:"source"`
 	SessionID   string `json:"session_id"`
-	Version     uint32 `json:"version"`
+	Version     int64  `json:"version"`
 }
 
 // DeltaJSON json model of a change record in a DeltaFile
